Update lightfm shim with codereview feedback
Also share code with the implicit shim
