+++ requirements/lightfm.txt
@@ -1,0 +1,1 @@
+lightfm>=1.0.0
