+++ tests/lightfm/test_lightfm.py
@@ -15,18 +15,15 @@
 #
 from merlin.io import Dataset
 from merlin.models.data.synthetic import SyntheticData
-from merlin.models.implicit import AlternatingLeastSquares
+from merlin.models.lightfm import LightFM
 from merlin.schema import Tags
 
 
-def test_alternating_least_squares(music_streaming_data: SyntheticData):
+def test_warp(music_streaming_data: SyntheticData):
     music_streaming_data._schema = music_streaming_data.schema.remove_by_tag(Tags.TARGET)
-
-    model = AlternatingLeastSquares(factors=128, iterations=15, regularization=0.01)
-
     dataset = Dataset(music_streaming_data.dataframe, schema=music_streaming_data.schema)
+
+    model = LightFM(learning_rate=0.05, loss="warp", epochs=10)
     model.fit(dataset)
 
-    metrics = model.evaluate(dataset)
-
-    assert all(metric >= 0 for metric in metrics.values())
+    model.predict(dataset)
